command_stop: add extra cast to make compiler happy
http://monetdb.cwi.nl/testing/projects/monetdb/testweb/logs/39420:80bbb901f820/GNU-Solaris-sparc/make.html#l2598

+++ tools/merovingian/daemon/argvcmds.c
@@ -346,7 +346,7 @@ command_stop(confkeyval *ckv, int argc, char *argv[])
 
 	if (kill(daemon, SIGTERM) == -1) {
 		fprintf(stderr, "unable to shut down monetdbd[%d]: %s\n",
-				daemon, strerror(errno));
+				(int)daemon, strerror(errno));
 		return(1);
 	}
